Begin switch to using dynamic containers for all of our computation.
Because of the buffer, I only intitialized the fist container with the BreakpointDetector constructor. The idea right now is to copy construct using the last element, and then go backwards through the list until we find the best element. This allows me to save the state of the last container before updating it while adding minimal operations.